authorization&authentication

+++ src/tasks/controller.py
@@ -2,20 +2,23 @@ from src.tasks.dtos import TaskSchema
 from sqlalchemy.orm import Session
 from src.tasks.models import TaskModel
 from fastapi import HTTPException
+from src.user.models import UserModel
 
-def create_task(body:TaskSchema, db:Session):
+
+def create_task(body:TaskSchema, db:Session, user:UserModel):
     data = body.model_dump()
     new_task = TaskModel(title = data["title"], 
                          description = data["description"], 
-                         is_completed = data["is_completed"]
+                         is_completed = data["is_completed"],
+                         user_id = user.id
                          )
     db.add(new_task)
     db.commit()
     db.refresh(new_task)
     return new_task
 
-def get_tasks(db:Session):
-    tasks = db.query(TaskModel).all()
+def get_tasks(db:Session, user:UserModel):
+    tasks = db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
     return tasks
 
 def get_by_id(id: int,db: Session):
@@ -24,10 +27,12 @@ def get_by_id(id: int,db: Session):
         raise HTTPException(404, detail="Id is incorrect")
     return one_task
 
-def update_task(body : TaskSchema, id : int, db : Session):
+def update_task(body : TaskSchema, id : int, db : Session, user:UserModel):
     one_task = db.query(TaskModel).get(id)
     if not one_task:
-        raise HTTPException(404, detail="Id is incorrect")
+        raise HTTPException(404, detail="Task Id is incorrect")
+    if one_task.user_id != user.id:
+        raise HTTPException(404, detail="You are not allowed to change this task..")
     
     body = body.model_dump()
     for field, value in body.items():
@@ -38,11 +43,12 @@ def update_task(body : TaskSchema, id : int, db : Session):
     db.refresh(one_task)
     return one_task
 
-def delete_task(id : int, db : Session):
+def delete_task(id : int, db : Session, user:UserModel):
     one_task = db.query(TaskModel).get(id)
     if not one_task:
         raise HTTPException(404, detail="Id is incorrect")
-    
+    if one_task.user_id != user.id:
+        raise HTTPException(404, detail="You are not authorized to delete this task..")
     db.delete(one_task)
     db.commit()
